rename CustomerInformation to ViewableCustomer and move related dto's to Viewable namespace
