+++ src/Core/Domain/Customer/Dto/Viewable/ViewedProductInformation.php
@@ -24,7 +24,7 @@
  * International Registered Trademark & Property of PrestaShop SA
  */
 
-namespace PrestaShop\PrestaShop\Core\Domain\Customer\Dto;
+namespace PrestaShop\PrestaShop\Core\Domain\Customer\Dto\Viewable;
 
 /**
  * Class ViewedProductInformation holds information about product that customer has viewed.
